src/requirements.txt
Cython>=0.29.14
numpy~=1.21
scipy==1.7.3
csm_openbabel==3.1
